+++ users/controllers/__init__.py
@@ -1,0 +1,1 @@
+from .signup_controller import SignupController
